Rewrite the API in a cooler way. Just drop the stupid repetition and use some dynamic swag.
