--- slack-client/__init__.py
@@ -1,3 +1,0 @@
-from .slack_api import SlackAPI
-
-__version__ = '0.1.0'
