+++ src/__init__.py
@@ -1,0 +1,3 @@
+from api import SlackAPI
+
+__version__ = '0.1.0'
